Add global data list

+++ ekikan.ml
@@ -3,6 +3,6 @@ type ekikan_t = {
   kiten : string;  (* 起点の駅名 *)
   shuten : string; (* 終点の駅名 *)
   keiyu : string;  (* 経由駅の駅名 *)
-  kyori : int;     (* 距離 *)
+  kyori : float;   (* 距離 *)
   jikan : int;     (* 所要時間 *)
 }
